Adde GO and NCBITaxonomy data bundles
