+++ src/main/scala/enzyme/bundles.scala
@@ -1,6 +1,6 @@
 package com.bio4j.data.enzyme
 
-import com.bio4j.data._
+import com.bio4j.data.bundles._
 import java.net.URL
 import better.files._
 
@@ -13,13 +13,13 @@ case object bundles {
     ).map { suffix =>
       new URL("ftp", "ftp.ebi.ac.uk", s"/pub/databases/enzyme/release/${suffix}")
     },
-    baseDirectory = file"/media/ephemeral0/enzyme/raw/",
+    baseDirectory = file"/media/ephemeral0/enzyme/data/",
     gunzip = false
   )()
 
   case object copyData extends CopyToS3(
     rawData.files,
-    ???
+    releasesPrefix / "data" / "enzyme" /
   )()
 
   case object mirroredData extends GetS3Copy(
